Issue #14769: Implemented blocks package structure for IT Regression area
