--- src/it/java/org/checkstyle/suppressionxpathfilter/XpathRegressionEmptyBlockTest.java
@@ -1,80 +1,0 @@
-///////////////////////////////////////////////////////////////////////////////////////////////
-// checkstyle: Checks Java source code and other text files for adherence to a set of rules.
-// Copyright (C) 2001-2025 the original author or authors.
-//
-// This library is free software; you can redistribute it and/or
-// modify it under the terms of the GNU Lesser General Public
-// License as published by the Free Software Foundation; either
-// version 2.1 of the License, or (at your option) any later version.
-//
-// This library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-// Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public
-// License along with this library; if not, write to the Free Software
-// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
-///////////////////////////////////////////////////////////////////////////////////////////////
-
-package org.checkstyle.suppressionxpathfilter;
-
-import java.io.File;
-import java.util.Collections;
-import java.util.List;
-
-import org.junit.jupiter.api.Test;
-
-import com.puppycrawl.tools.checkstyle.DefaultConfiguration;
-import com.puppycrawl.tools.checkstyle.checks.blocks.EmptyBlockCheck;
-
-public class XpathRegressionEmptyBlockTest extends AbstractXpathTestSupport {
-    private final String checkName = EmptyBlockCheck.class.getSimpleName();
-
-    @Override
-    protected String getCheckName() {
-        return checkName;
-    }
-
-    @Test
-    public void testEmptyForLoopEmptyBlock() throws Exception {
-        final File fileToProcess =
-                new File(getPath("InputXpathEmptyBlockEmpty.java"));
-        final DefaultConfiguration moduleConfig =
-                createModuleConfig(EmptyBlockCheck.class);
-        moduleConfig.addProperty("option", "TEXT");
-        final String[] expectedViolation = {
-            "5:38: " + getCheckMessage(EmptyBlockCheck.class,
-                EmptyBlockCheck.MSG_KEY_BLOCK_EMPTY, "for"),
-        };
-        final List<String> expectedXpathQueries = Collections.singletonList(
-                "/COMPILATION_UNIT"
-                    + "/CLASS_DEF[./IDENT[@text='InputXpathEmptyBlockEmpty']]"
-                    + "/OBJBLOCK/METHOD_DEF[./IDENT[@text='emptyLoop']]"
-                    + "/SLIST/LITERAL_FOR/SLIST"
-        );
-        runVerifications(moduleConfig, fileToProcess, expectedViolation,
-                expectedXpathQueries);
-    }
-
-    @Test
-    public void testEmptyForLoopEmptyStatement() throws Exception {
-        final File fileToProcess =
-                new File(getPath("InputXpathEmptyBlockEmpty.java"));
-        final DefaultConfiguration moduleConfig =
-                createModuleConfig(EmptyBlockCheck.class);
-        final String[] expectedViolation = {
-            "5:38: " + getCheckMessage(EmptyBlockCheck.class,
-                EmptyBlockCheck.MSG_KEY_BLOCK_NO_STATEMENT),
-        };
-        final List<String> expectedXpathQueries = Collections.singletonList(
-                "/COMPILATION_UNIT"
-                        + "/CLASS_DEF[./IDENT[@text='InputXpathEmptyBlockEmpty']]"
-                        + "/OBJBLOCK/METHOD_DEF[./IDENT[@text='emptyLoop']]"
-                        + "/SLIST/LITERAL_FOR/SLIST"
-        );
-        runVerifications(moduleConfig, fileToProcess, expectedViolation,
-                expectedXpathQueries);
-    }
-
-}
