--- src/it/resources/org/checkstyle/suppressionxpathfilter/avoidnestedblocks/InputXpathAvoidNestedBlocksEmpty.java
@@ -1,8 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.avoidnestedblocks;
-
-public class InputXpathAvoidNestedBlocksEmpty {
-
-    void empty() {
-        {} // warn
-    }
-}
