--- src/it/resources/org/checkstyle/suppressionxpathfilter/avoidnestedblocks/InputXpathAvoidNestedBlocksVariable.java
@@ -1,11 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.avoidnestedblocks;
-
-public class InputXpathAvoidNestedBlocksVariable {
-
-    void varAssign() {
-        int whichIsWhich = 0;
-        { // warn
-            whichIsWhich = 2;
-        }
-    }
-}
