+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/avoidnestedblocks/InputXpathAvoidNestedBlocksBreakOutside.java
@@ -1,0 +1,19 @@
+package org.checkstyle.suppressionxpathfilter.blocks.avoidnestedblocks;
+
+public class InputXpathAvoidNestedBlocksBreakOutside {
+    int s(int a) {
+        int x;
+        int y;
+        switch (a) {
+            case 0: { // warn: break outside block
+                x = 2;
+                y = -2;
+            }
+            break;
+            default:
+                x = 3;
+                y = -3;
+        }
+        return x + y;
+    }
+}
