+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/avoidnestedblocks/InputXpathAvoidNestedBlocksEmpty.java
@@ -1,0 +1,8 @@
+package org.checkstyle.suppressionxpathfilter.blocks.avoidnestedblocks;
+
+public class InputXpathAvoidNestedBlocksEmpty {
+
+    void empty() {
+        {} // warn
+    }
+}
