+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/avoidnestedblocks/InputXpathAvoidNestedBlocksNotAllowedInSwitchCase.java
@@ -1,0 +1,20 @@
+package org.checkstyle.suppressionxpathfilter.blocks.avoidnestedblocks;
+
+public class InputXpathAvoidNestedBlocksNotAllowedInSwitchCase {
+
+    int s(int a) {
+        int x;
+        int y;
+        switch (a) {
+            case 0: { // warn: allowInSwitchCase=false
+                x = 2;
+                y = -2;
+                break;
+            }
+            default:
+                x = 3;
+                y = -3;
+        }
+        return x + y;
+    }
+}
