+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/avoidnestedblocks/InputXpathAvoidNestedBlocksVariable.java
@@ -1,0 +1,11 @@
+package org.checkstyle.suppressionxpathfilter.blocks.avoidnestedblocks;
+
+public class InputXpathAvoidNestedBlocksVariable {
+
+    void varAssign() {
+        int whichIsWhich = 0;
+        { // warn
+            whichIsWhich = 2;
+        }
+    }
+}
