+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/emptycatchblock/InputXpathEmptyCatchBlockTwo.java
@@ -1,0 +1,11 @@
+package org.checkstyle.suppressionxpathfilter.blocks.emptycatchblock;
+
+public class InputXpathEmptyCatchBlockTwo {
+    public static void main(String[] args) {
+
+        try {
+            throw new RuntimeException();
+        } catch (RuntimeException e) /*warn*/ {
+        }
+    }
+}
