+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/leftcurly/InputXpathLeftCurlyOne.java
@@ -1,0 +1,5 @@
+package org.checkstyle.suppressionxpathfilter.blocks.leftcurly;
+
+public class InputXpathLeftCurlyOne
+{ //warn
+}
