+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/leftcurly/InputXpathLeftCurlyTwo.java
@@ -1,0 +1,4 @@
+package org.checkstyle.suppressionxpathfilter.blocks.leftcurly;
+
+public class InputXpathLeftCurlyTwo { //warn
+}
