+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/needbraces/InputXpathNeedBracesDo.java
@@ -1,0 +1,15 @@
+package org.checkstyle.suppressionxpathfilter.blocks.needbraces;
+
+public class InputXpathNeedBracesDo {
+    /** @return helper func **/
+    boolean condition()
+    {
+        return false;
+    }
+
+    /** Test do/while loops **/
+    public void test() {
+        // Invalid
+        do test(); while (condition()); // warn
+    }
+}
