+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/needbraces/InputXpathNeedBracesEmptyLoopBody.java
@@ -1,0 +1,11 @@
+package org.checkstyle.suppressionxpathfilter.blocks.needbraces;
+
+public class InputXpathNeedBracesEmptyLoopBody {
+    private int incrementValue() {
+        return 1;
+    }
+
+    public void test() {
+        while(incrementValue() < 5);; // warn
+    }
+}
