+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/rightcurly/InputXpathRightCurlyFour.java
@@ -1,0 +1,10 @@
+package org.checkstyle.suppressionxpathfilter.blocks.rightcurly;
+
+public class InputXpathRightCurlyFour {
+    public void sample(boolean flag) {
+        if (flag) {
+            System.identityHashCode("heh");
+            flag = !flag; } String.CASE_INSENSITIVE_ORDER. //warn
+                equals("Xe-xe");
+    }
+}
