+++ src/it/resources/org/checkstyle/suppressionxpathfilter/blocks/rightcurly/InputXpathRightCurlyThree.java
@@ -1,0 +1,7 @@
+package org.checkstyle.suppressionxpathfilter.blocks.rightcurly;
+
+public class InputXpathRightCurlyThree {
+    public void sample(boolean flag) {
+        if (flag) { String.CASE_INSENSITIVE_ORDER.equals("it is ok."); } //warn
+    }
+}
