--- src/it/resources/org/checkstyle/suppressionxpathfilter/emptyblock/InputXpathEmptyBlockEmpty.java
@@ -1,8 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.emptyblock;
-
-public class InputXpathEmptyBlockEmpty {
-    private void emptyLoop() {
-        for (int i = 0; i < 10; i++) {} // warn
-
-    }
-}
