--- src/it/resources/org/checkstyle/suppressionxpathfilter/emptycatchblock/InputXpathEmptyCatchBlockOne.java
@@ -1,10 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.emptycatchblock;
-
-public class InputXpathEmptyCatchBlockOne {
-    public static void main(String[] args) {
-
-        try {
-            throw new RuntimeException();
-        } catch (RuntimeException e) {} //warn
-    }
-}
