--- src/it/resources/org/checkstyle/suppressionxpathfilter/emptycatchblock/InputXpathEmptyCatchBlockTwo.java
@@ -1,11 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.emptycatchblock;
-
-public class InputXpathEmptyCatchBlockTwo {
-    public static void main(String[] args) {
-
-        try {
-            throw new RuntimeException();
-        } catch (RuntimeException e) /*warn*/ {
-        }
-    }
-}
