--- src/it/resources/org/checkstyle/suppressionxpathfilter/leftcurly/InputXpathLeftCurlyOne.java
@@ -1,5 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.leftcurly;
-
-public class InputXpathLeftCurlyOne
-{ //warn
-}
