--- src/it/resources/org/checkstyle/suppressionxpathfilter/leftcurly/InputXpathLeftCurlyTwo.java
@@ -1,4 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.leftcurly;
-
-public class InputXpathLeftCurlyTwo { //warn
-}
