--- src/it/resources/org/checkstyle/suppressionxpathfilter/needbraces/InputXpathNeedBracesDo.java
@@ -1,15 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.needbraces;
-
-public class InputXpathNeedBracesDo {
-    /** @return helper func **/
-    boolean condition()
-    {
-        return false;
-    }
-
-    /** Test do/while loops **/
-    public void test() {
-        // Invalid
-        do test(); while (condition()); // warn
-    }
-}
