--- src/it/resources/org/checkstyle/suppressionxpathfilter/needbraces/InputXpathNeedBracesEmptyLoopBody.java
@@ -1,11 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.needbraces;
-
-public class InputXpathNeedBracesEmptyLoopBody {
-    private int incrementValue() {
-        return 1;
-    }
-
-    public void test() {
-        while(incrementValue() < 5);; // warn
-    }
-}
