--- src/it/resources/org/checkstyle/suppressionxpathfilter/needbraces/InputXpathNeedBracesSingleLineLambda.java
@@ -1,6 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.needbraces;
-
-public class InputXpathNeedBracesSingleLineLambda {
-    static Runnable r3 = () -> // warn
-            String.CASE_INSENSITIVE_ORDER.equals("Hello world two!");
-}
