--- src/it/resources/org/checkstyle/suppressionxpathfilter/rightcurly/InputXpathRightCurlyFour.java
@@ -1,10 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.rightcurly;
-
-public class InputXpathRightCurlyFour {
-    public void sample(boolean flag) {
-        if (flag) {
-            System.identityHashCode("heh");
-            flag = !flag; } String.CASE_INSENSITIVE_ORDER. //warn
-                equals("Xe-xe");
-    }
-}
