--- src/it/resources/org/checkstyle/suppressionxpathfilter/rightcurly/InputXpathRightCurlyOne.java
@@ -1,13 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.rightcurly;
-
-public class InputXpathRightCurlyOne {
-    public void test(int x) {
-        if (x > 0)
-        {
-            return;
-        } //warn
-        else if (x < 0) {
-            ;
-        }
-    }
-}
