--- src/it/resources/org/checkstyle/suppressionxpathfilter/rightcurly/InputXpathRightCurlyThree.java
@@ -1,7 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.rightcurly;
-
-public class InputXpathRightCurlyThree {
-    public void sample(boolean flag) {
-        if (flag) { String.CASE_INSENSITIVE_ORDER.equals("it is ok."); } //warn
-    }
-}
