--- src/it/resources/org/checkstyle/suppressionxpathfilter/rightcurly/InputXpathRightCurlyTwo.java
@@ -1,11 +1,0 @@
-package org.checkstyle.suppressionxpathfilter.rightcurly;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-
-public class InputXpathRightCurlyTwo {
-    public void fooMethod() throws IOException {
-        try (BufferedReader br1 = new BufferedReader(null)) {
-            ; } //warn
-    }
-}
